Add DeleteSubcategoryOperation

+++ TransactionService/src/TransactionService/Domain/Services/Categories/UpdateCategoryOperations/DeleteSubcategoryOperation.cs
@@ -1,5 +1,6 @@
+using System.Collections.Generic;
+using System.Linq;
 using System.Threading.Tasks;
-using Microsoft.AspNetCore.JsonPatch.Operations;
 using TransactionService.Domain.Services.Categories.Exceptions;
 using TransactionService.Dtos;
 using TransactionService.Repositories;
@@ -9,13 +10,15 @@ namespace TransactionService.Domain.Services.Categories.UpdateCategoryOperations
     public class DeleteSubcategoryOperation : IUpdateCategoryOperation
     {
         private readonly ICategoriesRepository _categoriesRepository;
+        private readonly ITransactionHelperService _transactionHelperService;
         private readonly string _existingCategoryName;
         private readonly string _subcategoryName;
 
-        public DeleteSubcategoryOperation(ICategoriesRepository categoriesRepository, string existingCategoryName,
-            string subcategoryName)
+        public DeleteSubcategoryOperation(ICategoriesRepository categoriesRepository,
+            ITransactionHelperService transactionHelperService, string existingCategoryName, string subcategoryName)
         {
             _categoriesRepository = categoriesRepository;
+            _transactionHelperService = transactionHelperService;
             _existingCategoryName = existingCategoryName;
             _subcategoryName = subcategoryName;
         }
@@ -31,6 +34,16 @@ namespace TransactionService.Domain.Services.Categories.UpdateCategoryOperations
             if (!category.Subcategories.Contains(_subcategoryName))
                 throw new UpdateCategoryOperationException(
                     $"Failed to execute DeleteSubcategoryOperation, subcategory {_subcategoryName} under category {_existingCategoryName} does not exist");
+
+            var transactions = await _transactionHelperService.GetTransactionsAsync(new GetTransactionsQuery
+            {
+                Categories = new List<string> {_existingCategoryName},
+                Subcategories = new List<string> {_subcategoryName}
+            });
+
+            if (transactions.Any())
+                throw new UpdateCategoryOperationException(
+                    $"Failed to execute DeleteSubcategoryOperation, transactions in subcategory {_subcategoryName} still exist");
 
             await _categoriesRepository.DeleteSubcategory(_existingCategoryName, _subcategoryName);
         }
